# control-plane/rest/openapi-specs/v0_api_spec.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
  - url: /v0
paths:
...
  "/nodes/{node}/block_devices":
    get:
      tags:
        - BlockDevices
      operationId: get_node_block_devices
      parameters:
        - in: query
          name: all
          description: specifies whether to list all devices or only usable ones
          schema:
            type: boolean
        - in: path
          name: node
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockDevice"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BlockDevice:
      example:
        available: false
        devlinks:
          - ""
        devmajor: 0
        devminor: 0
        devname: ""
        devpath: ""
        devtype: ""
        filesystem:
          fstype: ""
          label: ""
          mountpoint: ""
          uuid: ""
        model: ""
        partition:
          name: ""
          number: 0
          parent: ""
          scheme: ""
          typeid: ""
          uuid: ""
        size: 0
      description: Block device information
      type: object
      properties:
        available:
          description: "identifies if device is available for use (ie. is not \"currently\" in\n use)"
          type: boolean
        devlinks:
          description: list of udev generated symlinks by which device may be identified
          type: array
          items:
            type: string
        devmajor:
          description: major device number
          type: integer
          format: int32
        devminor:
          description: minor device number
          type: integer
          format: int32
        devname:
          description: entry in /dev associated with device
          type: string
        devpath:
          description: official device path
          type: string
        devtype:
          description: "currently \"disk\" or \"partition\""
          type: string
        filesystem:
          example:
            fstype: ""
            label: ""
            mountpoint: ""
            uuid: ""
          description: filesystem information in case where a filesystem is present
          type: object
          properties:
            fstype:
              description: "filesystem type: ext3, ntfs, ..."
              type: string
            label:
              description: volume label
              type: string
            mountpoint:
              description: path where filesystem is currently mounted
              type: string
            uuid:
              description: UUID identifying the volume (filesystem)
              type: string
          required:
            - fstype
            - label
            - mountpoint
            - uuid
        model:
          description: device model - useful for identifying mayastor devices
          type: string
        partition:
          example:
            name: ""
            number: 0
            parent: ""
            scheme: ""
            typeid: ""
            uuid: ""
          description: partition information in case where device represents a partition
          type: object
          properties:
            name:
              description: partition name
              type: string
            number:
              description: partition number
              type: integer
              format: int32
            parent:
              description: devname of parent device to which this partition belongs
              type: string
            scheme:
              description: "partition scheme: gpt, dos, ..."
              type: string
            typeid:
              description: partition type identifier
              type: string
            uuid:
              description: UUID identifying partition
              type: string
          required:
            - name
            - number
            - parent
            - scheme
            - typeid
            - uuid
        size:
          description: size of device in (512 byte) blocks
          type: integer
          format: int64
      required:
        - available
        - devlinks
        - devmajor
        - devminor
        - devname
        - devpath
        - devtype
        - filesystem
        - model
        - partition
        - size
...
    RestJsonError:
      example:
        details: "The Pool 'pooloop' was not found"
        kind: NotFound
      description: Rest Json Error format
      type: object
      properties:
        details:
          description: detailed error information
          type: string
        kind:
          description: error kind
          type: string
          enum:
            - Timeout
            - Deserialize
            - Internal
            - InvalidArgument
            - DeadlineExceeded
            - NotFound
            - AlreadyExists
            - PermissionDenied
            - ResourceExhausted
            - FailedPrecondition
            - NotShared
            - NotPublished
            - AlreadyPublished
            - AlreadyShared
            - Aborted
            - OutOfRange
            - Unimplemented
            - Unavailable
            - Unauthenticated
            - Unauthorized
            - Conflict
            - FailedPersist
            - Deleting
      required:
        - details
        - kind
...